Clarifying the arguments in addAlias
I just added a sentence in the addAlias documentation to make it a bit clearer which string is intended to be used during ability checks when you create an alias. The code example makes sense, but our team was a little confused about which argument was really the alias.

I think the confusion lies in the usage of the term "alias". The way the code works, `update` and `delete` are actually aliases for `modify`, because they are the alternative names used in ability checks for the real defined rule, `modify`.